fix(navigation-instruction): fix the order of delayed swaps.
Previously, when a child route is loaded at first time, the swap() of child router-view is called before the swap() of parent router-view. The result is when binding the inner view, it could not get correct binding scope since its owningView was not bound yet. This fix makes sure the swaps are executed in the right order.

Closes #548